+++ grassfeed_rs/fr_core/src/ui_select/uimock.rs
@@ -130,10 +130,10 @@ impl UIAdapterValueStore for AdValSto {
     }
 
     fn set_tree_row_expand(&mut self, idx: usize, column: usize, bitmask: u32) {
-        while self.expand_data.len() < (idx + 1) as usize {
+        while self.expand_data.len() < (idx + 1) {
             self.expand_data.push((0, 0));
         }
-        self.expand_data[idx as usize] = (column as usize, bitmask);
+        self.expand_data[idx ] = (column, bitmask);
     }
     fn get_tree_row_expand(&self, idx: usize) -> (usize, u32) {
         self.expand_data[idx]
